Add systemTask constructor that accepts a WorkflowProcess, not just its variables for candidate groups in task list, cast ids to strings remove start process link from workflow summary page
