+++ src/org/labkey/workflow/view/workflowSummary.jsp
@@ -50,16 +50,6 @@
 <ul>
     <li>
         <%= PageFlowUtil.textLink("Active processes ", new ActionURL(WorkflowController.InstanceListAction.class, getContainer()).addParameter("processDefinitionKey", bean.getProcessDefinitionKey())) %>
-<%
-        if (bean.canStartProcess(getUser(), getContainer()))
-        {
-
-%>
-        &nbsp;&nbsp;
-        <%= PageFlowUtil.textLink("Start new process", new ActionURL(WorkflowController.StartProcessFormAction.class, getContainer()).addParameter("processDefinitionKey", bean.getProcessDefinitionKey()).addParameter("workflowModelModule", bean.getWorkflowModelModule())) %>
-<%
-        }
-%>
     </li>
 
     <li>Tasks</li>
